requirements.txt
autogen>=0.2.0
chromadb>=0.4.0
sentence-transformers>=2.2.2
pyyaml>=6.0
semgrep>=1.0.0
requests>=2.28.0
numpy>=1.21.0
torch>=2.0.0
transformers>=4.30.0
sentencepiece>=0.1.99
protobuf>=3.20.0